[network] Separate Peer and PeerHandle out from PeerManager
